% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_kernelshap.R
\name{aggregate_kernelshap}
\alias{aggregate_kernelshap}
\title{Aggregate SHAP objects}
\usage{
aggregate_kernelshap(x, ...)
}
\arguments{
\item{x}{a \code{kernelshap} object or a list of them.}

\item{...}{other \code{kernelshap} objects when \code{x} is not a list.}
}
\value{
A \code{kernelshap} object aggregating all the inputs.
}
\description{
Useful for reporting SHAP for multiple models in a cross validation schema.
}
\details{
SHAP values, data, standar errors, number of iterations, convergence and predictions of
each object are concatenated in the result. The other components of the results are
averaged or unique.
}
