% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.permshap}
\alias{is.permshap}
\title{Check for permshap}
\usage{
is.permshap(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
\code{TRUE} if \code{object} is of class "permshap", and \code{FALSE} otherwise.
}
\description{
Is object of class "permshap"?
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
s <- permshap(fit, iris[1:2, -1], bg_X = iris[, -1])
is.permshap(s)
is.permshap("a")
}
\seealso{
\code{\link[=kernelshap]{kernelshap()}}
}
