% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.permshap}
\alias{summary.permshap}
\title{Summarizes "permshap" Object}
\usage{
\method{summary}{permshap}(object, compact = FALSE, n = 2L, ...)
}
\arguments{
\item{object}{An object of class "permshap".}

\item{compact}{Set to \code{TRUE} for a more compact summary.}

\item{n}{Maximum number of rows of SHAP values etc. to print.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Summarizes "permshap" Object
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
s <- permshap(fit, iris[1:3, -1], bg_X = iris[, -1])
summary(s)
}
\seealso{
\code{\link[=permshap]{permshap()}}
}
