% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z_exact}
\alias{Z_exact}
\title{List of weighted on/off matrices Z up to p = 4 features}
\format{
A list of matrices
}
\usage{
Z_exact
}
\description{
List of weighted on/off matrices Z up to p = 4 features
}
\keyword{datasets}
